修复bug

+++ src/views/financialManagement/scrapList/zfTemp/index.vue
@@ -25,10 +25,10 @@
       </div>
     </Modal>
     <Modal v-model="cancelShow" width="600" :mask-closable="false" :closable="false" :title="types === 1 ? '作废收款单':'作废付款单'">
-      <Form  :model="dataApi" ref="formRef" :label-width="100" :rules="ruleInline" onkeydown="if(event.keyCode==13)return false;">
-      <FormItem label="作废备注：" prop="deleteRemark">
-        <Input type="text" v-model="dataApi.deleteRemark" style="width: 300px;" placeholder="请输入..."></Input>
-      </FormItem>
+      <Form :model="dataApi" ref="formRef" :label-width="100" :rules="ruleInline" onkeydown="if(event.keyCode==13)return false;">
+        <FormItem label="作废备注：" prop="deleteRemark">
+          <Input type="text" v-model="dataApi.deleteRemark" style="width: 300px;" placeholder="请输入..."></Input>
+        </FormItem>
       </Form>
       <div slot="footer">
         <Button @click="cancelHide">取消</Button>
@@ -45,12 +45,12 @@
 </template>
 
 <script>
-import {
-  dateformat
-} from '@/utils/filters.js'
-import detailModal from './detailModal'
+  import {
+    dateformat
+  } from '@/utils/filters.js'
+  import detailModal from './detailModal'
   export default {
-    components:{
+    components: {
       detailModal
     },
     props: {
@@ -77,15 +77,15 @@ import detailModal from './detailModal'
             trigger: 'change'
           }],
         },
-        show:false,
-        showImg:  '',
+        show: false,
+        showImg: '',
         cancelShow: false,
         detailShow: false,
         activeItem: {},
         list: [],
         totalCount: 0,
         dataValue: ['', ''],
-        tableHeader:[{
+        tableHeader: [{
           title: '作废单编号',
           key: 'cancelTicketId',
           minWidth: 200,
@@ -108,96 +108,96 @@ import detailModal from './detailModal'
               }, str),
             ]);
           }
-        },{
+        }, {
           title: '公司名称',
           key: 'customerName',
           minWidth: 200,
-          render: (h,params) =>{
-            let t = params.row.customerName !='' ? params.row.customerName : '无'
-            return h('span',t)
+          render: (h, params) => {
+            let t = params.row.customerName != '' ? params.row.customerName : '无'
+            return h('span', t)
           }
-        },{
+        }, {
           title: '客户银行账号',
           key: 'customerBankCardNo',
           minWidth: 200,
-          render: (h,params) =>{
-            let t = params.row.customerBankCardNo !='' ? params.row.customerBankCardNo : '无'
-            return h('span',t)
+          render: (h, params) => {
+            let t = params.row.customerBankCardNo != '' ? params.row.customerBankCardNo : '无'
+            return h('span', t)
           }
-        },{
+        }, {
           title: '供应商名称',
           key: 'buserNamebu',
           minWidth: 200,
-          render: (h,params) =>{
-            let t = params.row.buserName !='' ? params.row.buserName : '无'
-            return h('span',t)
+          render: (h, params) => {
+            let t = params.row.buserName != '' ? params.row.buserName : '无'
+            return h('span', t)
           }
-        },{
+        }, {
           title: '供应商银行账号',
           key: 'buserBankCardNo',
           minWidth: 200,
-          render: (h,params) =>{
-            let t = params.row.buserBankCardNo !='' ? params.row.buserBankCardNo : '无'
-            return h('span',t)
+          render: (h, params) => {
+            let t = params.row.buserBankCardNo != '' ? params.row.buserBankCardNo : '无'
+            return h('span', t)
           }
-        },{
-          title: this.types === 1 ? '进账金额':'出账金额',
+        }, {
+          title: this.types === 1 ? '进账金额' : '出账金额',
           key: 'amount',
           minWidth: 100
-        },{
-          title: this.types === 1 ? '进账时间':'出账时间',
+        }, {
+          title: this.types === 1 ? '进账时间' : '出账时间',
           key: 'inTime',
           minWidth: 150,
-          render: (h,params) =>{
-            return h('span',params.row.inTime != '' ? dateformat(params.row.inTime):'暂无')
+          render: (h, params) => {
+            return h('span', params.row.inTime != '' ? dateformat(params.row.inTime) : '暂无')
           }
-        },{
+        }, {
           title: '平台账号',
           key: 'bankCardNo',
           minWidth: 200
-        },{
+        }, {
           title: '银行账号流水号',
           key: 'bankTradeNo',
           minWidth: 200
-        },{
+        }, {
           title: '附件',
           key: 'fileAddress',
           minWidth: 100,
           render: (h, params) => {
             let fileAddress = params.row.fileAddress;
-            if(fileAddress != ''){
-            return h('div', [
-              h('Button', {
-                props: {
-                  type: 'warning',
-                  size: 'small'
-                },
-                style: {
-                  marginRight: '5px'
-                },
-                on: {
-                  click: () => {
-                    this.previewImg( params.row.fileAddress)
+            if (fileAddress != '') {
+              return h('div', [
+                h('Button', {
+                  props: {
+                    type: 'warning',
+                    size: 'small'
+                  },
+                  style: {
+                    marginRight: '5px'
+                  },
+                  on: {
+                    click: () => {
+                      this.previewImg(params.row.fileAddress)
+                    }
                   }
-                }
-              }, '查看'),
-            ]);
-            }else{
-              return h('span','暂无')
+                }, '查看'),
+              ]);
+            } else {
+              return h('span', '暂无')
             }
           }
-        },{
+        }, {
           title: '操作人',
           key: 'createUser',
           minWidth: 150
-        },{
+        }, {
           title: '操作时间',
           key: 'updateTime',
           minWidth: 150,
-          render: (h,params) =>{
-            return h('span',params.row.updateTime != '' ? dateformat(params.row.updateTime):'暂无')
+          render: (h, params) => {
+            return h('span', params.row.updateTime != '' ? dateformat(params.row.updateTime) : '暂无')
           }
-        },{
+        }, {
           title: '备注',
           key: 'remark',
           minWidth: 200
@@ -207,33 +207,53 @@ import detailModal from './detailModal'
           fixed: 'right',
           width: 140,
           render: (h, params) => {
-            return h('div', [
-              h('Button', {
-                props: {
-                  type: 'warning',
-                  size: 'small'
-                },
-                style: {
-                  marginRight: '5px'
-                },
-                on: {
-                  click: () => {
-                    this.goDetail( params.row)
+            let status = params.row.status;
+            if (status === 9) {
+              return h('div', [
+                h('Button', {
+                  props: {
+                    type: 'warning',
+                    size: 'small'
+                  },
+                  style: {
+                    marginRight: '5px'
+                  },
+                  on: {
+                    click: () => {
+                      this.goDetail(params.row)
+                    }
                   }
-                }
-              }, '详情'),
-              h('Button', {
-                props: {
-                  type: 'warning',
-                  size: 'small'
-                },
-                on: {
-                  click: () => {
-                    this.deleteItem(params.row)
+                }, '详情')
+              ]);
+            } else {
+              return h('div', [
+                h('Button', {
+                  props: {
+                    type: 'warning',
+                    size: 'small'
+                  },
+                  style: {
+                    marginRight: '5px'
+                  },
+                  on: {
+                    click: () => {
+                      this.goDetail(params.row)
+                    }
                   }
-                }
-              }, '作废')
-            ]);
+                }, '详情'),
+                h('Button', {
+                  props: {
+                    type: 'warning',
+                    size: 'small'
+                  },
+                  on: {
+                    click: () => {
+                      this.deleteItem(params.row)
+                    }
+                  }
+                }, '作废')
+              ]);
+            }
           }
         }]
       }
@@ -264,14 +284,14 @@ import detailModal from './detailModal'
     methods: {
       resetFilter() {
         this.pageApi = {
-          currentPage: 1,
-          pageSize: 10,
-          timeBeigin: '',
-          timeEnd: '',
-          companyName: '',
-          cancelTicketId: ''
-        },
-        this.dataValue = ['', '']
+            currentPage: 1,
+            pageSize: 10,
+            timeBeigin: '',
+            timeEnd: '',
+            companyName: '',
+            cancelTicketId: ''
+          },
+          this.dataValue = ['', '']
         this.getList(this.handleFilter)
       },
       getList(params) {
@@ -283,45 +303,45 @@ import detailModal from './detailModal'
           }
         })
       },
-      cancelHide(){
+      cancelHide() {
         this.cancelShow = false;
         this.dataApi.deleteRemark = '';
         this.$refs.formRef.resetFields();
       },
-      changePage(data){
+      changePage(data) {
         this.dataApi.currentPage = page;
         this.getList(this.handleFilter)
       },
-      previewImg(data){
+      previewImg(data) {
         this.show = true;
         this.showImg = data;
       },
-      salesDetail(data){
+      salesDetail(data) {
         this.$router.push('./scrapDetail/' + data.cancelTicketId)
       },
-      goDetail(data){
+      goDetail(data) {
         this.activeItem = data || {};
         this.detailShow = true;
       },
-      deleteItem(data){
+      deleteItem(data) {
         this.cancelShow = true;
         this.dataApi.id = data.id;
       },
-      handleAction(){
+      handleAction() {
         this.$refs.formRef.validate((valid) => {
-          if(valid){
+          if (valid) {
             let saveApi = this.types === 1 ? this.api.receiptCancel : this.api.paymentCancel;
             let params = this.$clearData(this.dataApi)
-            this.$http.post(saveApi,params).then(res =>{
-              if(res.code === 1000){
+            this.$http.post(saveApi, params).then(res => {
+              if (res.code === 1000) {
                 this.$Message.success('作废成功');
                 this.cancelShow = false;
                 this.getList(this.handleFilter);
-              }else{
+              } else {
                 this.$Message.error(res.message);
               }
             })
-          }else{
+          } else {
             this.$Message.error('备注不能为空')
           }
         })
@@ -334,8 +354,8 @@ import detailModal from './detailModal'
 </script>
 
 <style lang='less' scoped>
-  .showImg{
-    img{
+  .showImg {
+    img {
       display: block;
       margin: 0 auto;
       max-width: 100%;
